# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP EDU — API de gestión educativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: Listar alumnos
      parameters:
        - in: query
          name: groupId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Lista de alumnos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Crear alumno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Alumno creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/bulk:
    post:
      operationId: createStudentsBulk
      tags: [students]
      summary: Alta masiva de alumnos (desde CSV)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentBulkInput"
      responses:
        "200":
          description: Resultado de la importación masiva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Obtener alumno
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alumno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Actualizar alumno
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Alumno actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Eliminar alumno
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Teacher ──────────────────────────────────────────
...
    Student:
      type: object
      required: [id, firstName, lastName, email, username, groupId, groupName, dolibarrSyncStatus, createdAt]
      properties:
        id:                  { type: integer }
        firstName:           { type: string }
        lastName:            { type: string }
        email:               { type: string }
        username:            { type: string }
        groupId:             { type: integer }
        groupName:           { type: string }
        dolibarrSyncStatus:  { type: string, enum: [pending, synced, error] }
        dolibarrSyncError:   { type: ["string", "null"] }
        dolibarrPassword:    { type: ["string", "null"], description: "Contraseña admin del Dolibarr propio del alumno." }
        companyName:         { type: ["string", "null"] }
        createdAt:           { type: string, format: date-time }
    StudentInput:
      type: object
      required: [firstName, lastName, email, username, password, groupId]
      properties:
        firstName:   { type: string, minLength: 1 }
        lastName:    { type: string, minLength: 1 }
        email:       { type: string, format: email }
        username:    { type: string, minLength: 3 }
        password:    { type: string, minLength: 6 }
        groupId:     { type: integer }
        companyName: { type: string }
    StudentUpdate:
      type: object
      properties:
        firstName:   { type: string, minLength: 1 }
        lastName:    { type: string, minLength: 1 }
        email:       { type: string, format: email }
        groupId:     { type: integer }
        companyName: { type: string }
    StudentBulkInput:
      type: object
      required: [groupId, students]
      properties:
        groupId:  { type: integer }
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentInput"
    BulkImportResult:
      type: object
      required: [created, errors]
      properties:
        created:  { type: integer }
        errors:
          type: array
          items:
            type: object
            required: [username, error]
            properties:
              username: { type: string }
              error:    { type: string }
...